#trivial Separate out createSigner function
