+++ packages/wallet-sdk/src/sign/util.ts
@@ -1,7 +1,11 @@
+import { StateUpdateListener } from './interface';
+import { SCWSigner } from './scw/SCWSigner';
 import { WalletLinkSigner } from './walletlink/WalletLinkSigner';
 import { Communicator } from ':core/communicator/Communicator';
+import { standardErrors } from ':core/error';
 import { ConfigMessage, MessageID, SignerType } from ':core/message';
-import { AppMetadata, Preference } from ':core/provider/interface';
+import { AppMetadata, Preference, Signer } from ':core/provider/interface';
+import { getCoinbaseInjectedSigner } from ':util/provider';
 import { ScopedLocalStorage } from ':util/ScopedLocalStorage';
 
 const SIGNER_TYPE_KEY = 'SignerType';
@@ -30,6 +34,35 @@ export async function fetchSignerType(params: {
   };
   const { data } = await communicator.postRequestAndWaitForResponse(request);
   return data as SignerType;
+}
+
+export function createSigner(params: {
+  signerType: SignerType;
+  metadata: AppMetadata;
+  communicator: Communicator;
+  updateListener: StateUpdateListener;
+}): Signer {
+  const { signerType, metadata, communicator, updateListener } = params;
+  switch (signerType) {
+    case 'scw':
+      return new SCWSigner({
+        metadata,
+        updateListener,
+        postMessageToPopup: communicator.postRequestAndWaitForResponse,
+      });
+    case 'walletlink':
+      return new WalletLinkSigner({
+        metadata,
+        updateListener,
+      });
+    case 'extension': {
+      const injectedSigner = getCoinbaseInjectedSigner();
+      if (!injectedSigner) {
+        throw standardErrors.rpc.internal('injected signer not found');
+      }
+      return injectedSigner;
+    }
+  }
 }
 
 async function listenForWalletLinkSessionRequest(
